Get rid of gcc warnings.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@6005 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CMS3DMeshFileLoader.cpp
@@ -193,7 +193,7 @@ bool CMS3DMeshFileLoader::load(io::IReadFile* file)
 
 	u8* buffer = new u8[fileSize];
 	size_t read = file->read(buffer, fileSize);
-	if (read != fileSize)
+	if (read != (size_t)fileSize)
 	{
 		delete [] buffer;
 		os::Printer::log("Could not read full file. Loading failed", file->getFileName(), ELL_ERROR);
